+++ Gruntfile.js
@@ -380,11 +380,11 @@ module.exports = function (grunt) {
 
         var git_tasks = [
             'exec:git_commit:Release v' + newVersion,
-            //'exec:git_flow_start:' + newVersion,
-            //'exec:git_flow_finish:' + newVersion + ':Release v' + newVersion,
-            //'exec:git_push:origin:develop:tags',
-            //'exec:git_push:origin:master:tags',
-            //'exec:pypi_publish'
+            'exec:git_flow_start:' + newVersion,
+            'exec:git_flow_finish:' + newVersion + ':Release v' + newVersion,
+            'exec:git_push:origin:develop:tags',
+            'exec:git_push:origin:master:tags',
+            'exec:pypi_publish'
         ];
 
         grunt.task.run(git_tasks);